Preparation for proper texture creation
We want to create textures once, and just reinit their image size when
we resize the window. I'm preparing to do that, but it will require some
major refactoring to make it even close to understandable.

The first step is to use some dedicated create and delete methods, so
that structs and their destruction are close together. This requires
some separate files, since keeping anything close together is impossible
when you file is >1K lines

The next step is to make some per window glx init function so we can
attach arbitrary shit to it. I'm thinking about just removing the
glx/nonglx abstraction. I'd like to make it more game like, and separate
all the drawing from the updating. We'll see.