Warning: large commit
- Move xsync over to the new X11Capabilities struct, and use it
everywhere.
- Remove some more superfluous properties of the session
- Add some new awesome capabilities to vector (like allocating space to
mess about with)
- Move rectangle triangulation over to face.c. That's really not an
ideal place, but that file is still small enough for it to be
acceptable.
- Make the face fetching a little more data oriented (by splitting to to
two steps)
- Allow faces to not be uploaded to the GPU, will be useful in testing
later
- Move face over to vector. Since we have it why not use it. The only
drawback is that all faces now cost a couple of extra size_t's, but
I can live with that. It also allows them to be dynamically sized, so
technically you shouldn't need to specify the mesh size in the .face
files anymore (you still do for now).
- A bunch more cleanup. The code is starting to be readable. All in all
a productive day.