500 lines of tests
I wanted to start testing the things that actually are testable.
I decided to start out with making a small testing framework and testing
my vector implementation in that. So far it's pretty neat.

The testing framework is based around the convention of a single assert
per test, and having the test defined by "an X will Y when Z". I find
this to be a neat way of organizing the tests, and it works pretty well
so far.

I'm currently collecting tests into a vector after returning, but it
should be relatively simple write them to shared memory and then later
append them if we decide to do some fork protection.

I'm not super happy with how the assert functionality works. It's sort
of nasty that you have to defined a struct for every type, and then an
enum, AND THEN a formatting message, in 3 completely separate places. On
the other side, it's neat to have all the strings isolated (because we
don't want strings everywhere);

I think this has potential.