Fork protect the tests
This should protect tests against memory corruption and asserts failing
the whole test suite. There's a fair number of hacks in here (honestly,
the IPC seems rather fragile), but I've played with it a bit and it
seems working. I should probably come around to fixing it later, but it
just doesn't seem worth the effort right now.