Bind a stencilbuffer on upscale
As it turns out, even if you aren't using the depth buffer (disable
depth_test and write) you still need to set it, because you can't unbind
it. We care because having a depth buffer smaller than the target
texture, even if you aren't using it, means we can' render to the whole
texture.