Add support for transparent entities
I've also made windows with a non-0 opacity transparent, meaning the
background is now also transparent. I don't know if this is the best
solution yet (users might not like it), but it's an easy fix for some of
the previous fading problems.

If I want to differentiate (so have background stay opaque on unfocus)
I'll have to add a second fader.