Selective redirection
First attempt at supporting selective redirection, allowing some
applications to avoid compositing.

There's some limitations. For one, I'm currently only checking for the
atom when mapping the window, so the property has to be set when the
window is mapped. There's also some pretty severe artifacts when mapping
and unmapping those windows.

More directly, I window that has opted to not be composited will never
be obscured, even if another window is on top of it.

It's pretty neat, but requires a lot more work