Remove unredir stuff and unredirected holepunching
Since we can now do selective redirection, we don't need global
unredirection, so we can remove that.

For some reason the compositor holepunching was a pixel off, so i fixed
that. I was also still adding the textured component to windows that
weren't redirected, causing us to render with an undefined texture.

I'd say selective redirection is kinda production ready now. I'd like to
have stacking working at some point, but it's the best implementation of
unredirection I've seen.