Actually render in pre-multiplied alpha
Most applications operate under the assumption that the compositors want
premultiplied alphas. Up until recently, we weren't most compositors and
actually wanted conventional alpha. That was mostly just down to me not
knowing how to do premultiplied alpha. Well, I learned and now we are
doing pre-* alpha as well.

The colors are so much more vibrant (and the fading actually works)!