Only fetch the information from fbconfig once
There's no reason to fetch the data from the FBConfig every time we bind
the drawable. Since binding the textures is BY FAR the most expensive
operation we do, we want to optimize that. This shaves a good amount of
time off it.

There might be more we can do, but this is a start.