Damage the blur of all windows above moving window
There was a bug where a window would only damage windows above the new
location when it moves (or resizes). Obviously, this can lead to
problems where windows we are no longer behind wont be updated to
reflect that fact.

I've also added a few performance notes that we'll have to look at

PS: This is a spooky commit