Fix fast switch window hang
A longstanding bug meant that after a fast switch, the some windows
might hang, and no longer get damaged.

To fix this I added a new debug view which prints all the components
currently attached to a window (given that at least physical is there as
well). It's super handy.

Turns out that i wasn't subtracting damage when the window was unmapped,
which meant that we didn't receive any new damage.

There's still a slight issue, where mapping and unmapping in the same
frame causes the system to get out of sync with the XServer. It should
be fixable with a bit more work, but for now you can at least fix it by
just unmapping and remapping again.