Make windowtype affect focus again
Due to the new debug system, we wanted to move all the component clears
down to the end of the rendering. That should have been fine, but it
turns out we had an implicit one frame delay when updating the window
type, because that was set after it was used. That meant that we were
now setting it just to clear it, not ideal. This just moved the set up
front, so we can actually use it on the frame where we need it.