Add separate bg opacity fade time
This can look nice in certain circumstances

They are synchronized at the starting oint, and just end when they will
(if foreground is 1000 and bg is 500, they start at the same time, which
means the bg ends before the foreground). I tried staggering them, so
that they would end at the same time when fading down and start at the
same time when fading up, but I couldn't find an elegant way of doing
it, and it didn't look good.

If the bg fade is longer than the foreground fade, you might get strange
results.