Split shadow stencil from image
In time I'd like to let shadow move behind their windows. To do so, we
will have to create the stencil where the window will be, while copying
the color data with some offset.

We still have to bind the stencil because otherwise opengl wont render
to the entire texture.

At the same time I noticed that we were doing some useless shader
rebinding, which i refactored out.