Split stencil (#15)
* Reuse local shadow border when resizing

Instead of resetting the border size when resizing the shadow region, we
might as well just reuse the one we already have.

* Split shadow stencil from image

In time I'd like to let shadow move behind their windows. To do so, we
will have to create the stencil where the window will be, while copying
the color data with some offset.

We still have to bind the stencil because otherwise opengl wont render
to the entire texture.

At the same time I noticed that we were doing some useless shader
rebinding, which i refactored out.

* Add shadow component debug

I'm having some issues with shadows on first draw. I've added
a component that draws the shadow texture.