Add hotloading of assets.
I've added a new hotloading system based on inotify. Basically, when
loading an asset we add the filepath to inotify, and then reload it
whenever the file changes.

Here we also keep track of what assets are loaded as part of loading
another asset, and add a dependency on those assets. We can then reload
the downstream assets when the upstream changes.

As it stands, asset users are not told their assets have reloaded.