Use square face when blitting shadow (#19)
When we are blitting the shadow (to stencil out the window or to render
it from cache) we were previously using the per window face. That
doesn't work if the window has a non-square shape since that will cut
shadow. Instead we can just blit it with the static square face, since
it's already stenciled out and sized correctly.

That's exactly what this change does. Shadows on Xeyes (which have been
broken for a long time) now render correctly.