Split out some of the functionality to systems
The blur and shadow workloads are already pretty separated from the rest
of the logic. Move that into some system files to make some smaller and
more clear files. This also made it obvious that the blur uses some
temporary vectors that might as well be allocated at startup and reused
between frames. less malloc, yay.