Discard background
When blitting the (blurred) background to the screen there's no need to
render it where the image is going to completely obscure it anyway. We
can save some fillrate by discarding those fragments that are going to
lie behind opaque windows. Since there doesn't seem to be a nice way of
skipping the drawing of backgrounds behind opaque windows this is
probably the best we can do.

It's not clear to me if the additional read from another texture is more
expensive than the texture write, but users are free to rewrite the
shader if it is.