Track client-frame relationship in xorg system
Since i don't run a reparenting WM this required me to implement a ton
of tests. The xorg system is therfore (even at this stage where it's not
even half done) now the most well tested part of the system.

There's a lot of work to make the client->frame translation efficient
when we start dynamically translating events, but for now the
relationship is kept indirect.