Better handling of unshaped alpha blended windows
Firefox no longer does any attempt at doing shaped windows for ogl
composited windows. Since it's pretty clear that they don't consider it
a bug, i guess we have to handle alpha blended windows, even if they
don't inform the WM of their extents.

To do so i've refactored a bit of the shadow and background handling to
move the clipping out to a new blit shader that is also aware of the
window content texture (so that it can check the alpha value of the
pixel in the source). With that we can clip the blurred background and
the shadow wherever we want. For now i'm only clipping on 0 alpha.

This still has some visual artifacts for when firefox does
anti-aliasing, but I think it's the best I can do without knowing the
actual shape