Fix the old lazy rendering path
We used to only rerender the screen when there was a new X event that
might cause something to change. I dropped that feature some time ago by
mistake, since I disable to for testing. I've done a bit of work to add
it back, and also make it a bit more understandable for me.

There's still a bug that if we process some events, and then get an
event that gets filtered out by the xorg module, we will actually do
a rerender even if there's more events waiting. Ideally it should
process all the events before it tries to render any of them.