Allow for parallel make
With parallel make, we didn't ensure that the generated files were
generated before running the default target. That's a compile error.

I could fix it for real, but that seems hard, so for now I've created
a target _gen_ that you will have to run before starting a parallel make