Keep textures for windows that fail to bind
If a texture is updated right before it's unmapped, we might try to bind
it and get a failure. Before we just removed the texture in that case,
but such a window will still have a shadow and blur. Since we have an
expectation in those systems that a window always has a texture if it
has one of those, we have to keep the texture around.

It might be nicer to just remove that expectation, but it's too much for
work now. This fixes the crashbug