Hoist bypass detection into xorg translation layer
Handling of the bypass property is not inside the xorg layer. This means
it has tests and a clear interface to the core compositor. It also
means a window now as a new state.

Before a window could be either mapped or unmapped. Now it has a third
state, called _Bypassed_. A bypassed window is a mapped window that has
asked the compositor not to be composited. An unmapped window that has
the bypass property set is treated like any other unmapped window, but
when such a window gets mapped again, it's mapped into the bypassed
state.