Only emit bypass if the window is mapped
If the window isn't mapped we don't really care if it's bypassed. The
xorg input system still needs to keep track to tell us when it
eventually does get mapped. I wrote a couple of tests for that before
(when I broke the test system). This completes those tests too.