Add support for resizing canvas to core
It turns out the implementation was still there from back when configure
was the branching point, it was just never called because it was
filtered out.

I've removed support for reredir on resize. If that's required it will
be added without a configuration option. I don't see why the user should
have to configure something so specific.