UnBypass window when bypassed client is destroyed
If a window has a client which is bypassed, and that client is
destroyed we need to remap the client to make sure it isn't bypassed
anymore.

I've added a testcase for this, since I thought it already worked.
Unfortunately it actually didn't work. So i fixed it as well.