Fix flickering when bg fades longer than window
The transparent pass stopped rendering the window when it was done
fading, even though it can't be considered completely opaque before the
background is done fading. This mean that neither path drew the window
contents while the background was fading in.

I've added a bit of code to make sure the transparent pass draws the
window even if the contents are solid (since the contents is an ARGB
image we can't know if it's actually opaque).