Add rudimentary XResource tracking support
The (underdocumented) XResource interface lets us see how many resources
we have allocated on the XServer side. This should be pretty useful for
tracking leaks, but also looks like a nice optimization target. I've
added support for it in the framerate display interface. It currently
just shows the currently active resources (as observed by the XServer),
but it would be nice with some graphs as well.

The interface is handled by a single function in the xorg backend that
fills out two arrays, one with the names of the resources, and one with
the counts. The index of the resources is stable. The lookup from atom
to index is handled by the backend.

The name array is currently a little fragile, in that it may be
reallocated between frames, and so the debug renderer should not cache
the pointer. Also notice that the length of the raw name array is always
the same as the value vector.